+++ src/main/java/HomeworkImpl.java
@@ -3,10 +3,8 @@ import creators.translation.TranslationCreator;
 import creators.translation.TranslationCreatorImpl;
 import dao.DaoImpl;
 import dao.RadioStationDao;
-import dao.lexer.Converter;
-import dao.lexer.ConverterImpl;
+import dao.coverter.Converter;
 import dao.lexer.Lexer;
-import dao.lexer.LexerImpl;
 import lombok.SneakyThrows;
 import personality.Broadcaster;
 import personality.GuestBroadcaster;
