+++ src/test/java/dao/RadioStationDaoTest.java
@@ -1,6 +1,6 @@
 package dao;
 
-import dao.lexer.ConverterImpl;
+import dao.coverter.ConverterImpl;
 import dao.lexer.LexerImpl;
 import org.junit.jupiter.api.Test;
 import personality.RadioBroadcaster;
